Model {Value,Asset,Nonce}::Confidential using secp256k1-zkp types
This patch introduces a _breaking_ change in the (de)serialisation of
confidential assets, confidential values and confidential nonces. In
particular, we now support both readable and compact serialization,
and they both include a serialization prefix `2` as opposed to the old
prefix which is now part of the inner condifential value, since that
is controlled by secp256k1-zkp. We choose the number `2` since we were
already using `0` for null values and `1` for explicit values.

We also had to modify some of the test data now that `secp256k1-zkp` is
validating for us that the confidential assets are actually valid
generators.